Extension Methods Fix (#46)
Extension methods on HttpResponseMessage would attempt to read content from the Content property based on whether or not the status code on the response indicated a success. This was a failure for two reasons:

1. A successful response might not contain any content
2. A failure response might contain content about the failure

The methods have been refactored to key off of whether the Content property is not null.